+++ externals/coda-oss/modules/c++/coda_oss/include/coda_oss/string.h
@@ -37,9 +37,8 @@ namespace coda_oss
     #pragma warning(disable: 5052) // Keyword '...' was introduced in C++20 and requires use of the '...' command-line option
     #endif
     enum class char8_t : unsigned char { }; // https://en.cppreference.com/w/cpp/language/types
-    CODA_OSS_disable_warning_pop
-
     using u8string = std::basic_string<char8_t>; // https://en.cppreference.com/w/cpp/string
+    CODA_OSS_disable_warning_pop
 }
 
 #endif  // CODA_OSS_coda_oss_string_h_INCLUDED_
